Rename search modal to match convention
