+++ apps/frontend/src/app/shared/modals/search/index.ts
@@ -1,0 +1,1 @@
+export * from './search-modal';
