+++ apps/frontend/src/app/shared/modals/search/search-modal.module.scss
